+++ alpasso/src/main/scala/alpasso/service/fs/model/RawSecretData.scala
@@ -1,5 +1,7 @@
 package alpasso.service.fs.model
 
+import java.nio.file.Path
+
 import cats.Show
 
 opaque type RawSecretData = Array[Byte]
