+++ alpasso/src/main/scala/alpasso/service/fs/model/RepositoryConfiguration.scala
@@ -1,0 +1,8 @@
+package alpasso.service.fs.model
+
+import java.nio.file.Path
+
+import alpasso.common.SemVer
+import alpasso.service.cypher.CypherAlg
+
+case class RepositoryConfiguration(repoDir: Path, version: SemVer, cypherAlg: CypherAlg)
