+++ core/src/main/scala/alpasso/core/model/SecretPacket.scala
@@ -5,7 +5,7 @@ import cats.*
 case class SecretPacket[+T](name: SecretName, payload: T)
 
 object SecretPacket:
-  given Functor[SecretPacket] = new Functor[SecretPacket]:
+  given Functor[SecretPacket] with 
     def map[A, B](fa: SecretPacket[A])(f: A => B): SecretPacket[B] =
       SecretPacket(fa.name, f(fa.payload))
